interfaces/sloth/pyproject.toml
[project]
name = "charmlibs-interfaces-sloth"
description = "Sloth integration for Juju charms - SLO (Service Level Objective) management with the Sloth operator."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name="The Observability team at Canonical Ltd."},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 3 - Alpha",
]
dynamic = ["version"]
dependencies = [
    "ops",
    "pydantic>=2.0",
    "pyyaml",
    "cosl>=0.0.10",
]

[dependency-groups]
lint = [  # installed for `just lint interfaces/sloth` (unit, functional, and integration are also installed)
    "typing_extensions",
    "cosl>=0.0.10",
]
unit = [  # installed for `just unit interfaces/sloth`
    "ops[testing]",
    "pytest",
    "pytest-cov",
    "cosl>=0.0.10",
]
functional = [  # installed for `just functional interfaces/sloth`
]
integration = [  # installed for `just integration interfaces/sloth`
    "jubilant",
    "pytest-jubilant",
]

[project.urls]
"Repository" = "https://github.com/canonical/charmlibs"
"Issues" = "https://github.com/canonical/charmlibs/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/charmlibs"]

[tool.hatch.version]
path = "src/charmlibs/interfaces/sloth/_version.py"

[tool.ruff]
extend = "../../pyproject.toml"
src = ["src", "tests/unit", "tests/functional", "tests/integration"]  # correctly sort local imports in tests
exclude = [
    "tests/integration/charms/*/library",
]

[tool.ruff.lint.extend-per-file-ignores]
# add additional per-file-ignores here to avoid overriding repo-level config
"tests/**/*" = [
    # "E501",  # line too long
]

[tool.pyright]
extends = "../../pyproject.toml"
include = ["src", "tests"]
pythonVersion = "3.8"  # check no python > 3.8 features are used

[tool.charmlibs.functional]
ubuntu = []  # ubuntu versions to run functional tests with, e.g. "24.04" (defaults to just "latest")
pebble = []  # pebble versions to run functional tests with, e.g. "v1.0.0", "master" (defaults to no pebble versions)
sudo = false  # whether to run functional tests with sudo (defaults to false)

[tool.charmlibs.integration]
# tags to run integration tests with (defaults to running once with no tag, i.e. tags = [''])
# Available in CI in tests/integration/pack.sh and integration tests as CHARMLIBS_TAG
tags = []  # Not used by the pack.sh and integration tests generated by the template
